pyproject.toml
[project]
name = "shiftd"
version = "0.1.0"
description = "Any data to any data. Universal format converter."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.13"
keywords = ["data", "converter", "csv", "json", "xml", "pydantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
arrow = ["pyarrow>=15.0.0"]
excel = ["openpyxl>=3.1.0"]
yaml = ["pyyaml>=6.0.0"]
postgres = ["psycopg2-binary>=2.9.0"]
duckdb = ["duckdb>=1.0.0"]
mysql = ["pymysql>=1.1.0"]
llm-openai = ["openai>=1.0.0"]
llm-anthropic = ["anthropic>=0.18.0"]
llm-ollama = ["ollama>=0.3.0"]
all = ["shiftd[arrow,excel,yaml,postgres,duckdb,mysql,llm-openai,llm-anthropic,llm-ollama]"]

[project.scripts]
shiftd = "shiftd.cli:main"

[project.urls]
Repository = "https://github.com/datapoints-np/shiftd"
Documentation = "https://github.com/datapoints-np/shiftd#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shiftd"]

[tool.hatch.build.targets.sdist]
include = ["/shiftd"]

[dependency-groups]
dev = [
    "invoke>=2.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["shiftd", "tasks"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
